api/requirements.txt
flask
pandas
numpy
pyyaml
scikit-learn
joblib
openpyxl
